Fixed some docs, moved the github stuff into a single package, saved the run config
