+++ backend/src/main/java/io/github/orionhealth/xbdd/config/ResourceServerConfiguration.java
@@ -11,6 +11,7 @@
 package io.github.orionhealth.xbdd.config;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
@@ -24,6 +25,10 @@ public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter
 	public void configure(final HttpSecurity http) throws Exception {
 		http.authorizeRequests()
 				.antMatchers("/rest/attachment/**")
+				.permitAll()
+				.antMatchers(HttpMethod.PUT, "/rest/reports/**")
+				.permitAll()
+				.antMatchers(HttpMethod.POST, "/rest/reports/**")
 				.permitAll()
 				.antMatchers("/rest/**")
 				.authenticated();
